+++ app/controllers/collection_objects_controller.rb
@@ -443,12 +443,13 @@ class CollectionObjectsController < ApplicationController
       :never_loaned,
       :on_loan,
       :partial_overlap_dates,
-      :radius,
+      :preparation_type_id,
+      :radius,  # CE filter
       :repository,
       :repository_id,
       :sled_image_id,
       :spatial_geographic_areas,
-      :start_date,
+      :start_date,  # CE filter
       :taxon_determinations,
       :type_material,
       :type_specimen_taxon_name_id,
@@ -467,7 +468,8 @@ class CollectionObjectsController < ApplicationController
       keyword_id_and: [],
       keyword_id_or: [],
       otu_ids: [],
-      collector_ids: [] # 
+      collector_ids: [], #
+      preparation_type_id: []
       #  user_id: []
       #  collecting_event: {
       #   :recent,
@@ -496,7 +498,7 @@ class CollectionObjectsController < ApplicationController
       :depictions,
       :determiner_id_or,
       :dwc_indexed,
-      :end_date,
+      :end_date,  # CE filter
       :exact_buffered_collecting_event,
       :exact_buffered_determinations,
       :exact_buffered_other_labels,
@@ -511,13 +513,14 @@ class CollectionObjectsController < ApplicationController
       :in_labels,
       :in_verbatim_locality,
       :loaned,
-      :md5_verbatim_label,
+      :md5_verbatim_label, # CE filter
       :namespace_id,
       :never_loaned,
       :on_loan,
-      :partial_overlap_dates,
+      :partial_overlap_dates, # CE filter
       :radius,
       :repository,
+      :preparation_type_id,
       :repository_id,
       :sled_image_id,
       :spatial_geographic_areas,
@@ -530,17 +533,18 @@ class CollectionObjectsController < ApplicationController
       :user_id,
       :user_target,
       :validity,
-      :wkt,
+      :wkt, # CE filter
       biocuration_class_ids: [],
       biological_relationship_ids: [],
       collecting_event_ids: [],
+      collector_ids: [],
       determiner_id: [],
-      geographic_area_ids: [],
+      geographic_area_ids: [], # CE filter
       is_type: [],
       keyword_id_and: [],
       keyword_id_or: [],
       otu_ids: [],
-      collector_ids: []
+      preparation_type_id: [],
 
       #  collecting_event: {
       #   :recent,